// src/api/__generated__/wallet/model/followListItemResponseRelationStatus.ts
/**
 * Generated by orval v8.22.0 🍺
 * Do not edit manually.
 * Mini Drama User Wallet API
 * 用户、钱包、任务中心接口
 * OpenAPI spec version: v1.0
 */

/**
 * 当前登录用户对该项用户的关系状态（未登录为 NONE）
 */
export type FollowListItemResponseRelationStatus =
  (typeof FollowListItemResponseRelationStatus)[keyof typeof FollowListItemResponseRelationStatus];

export const FollowListItemResponseRelationStatus = {
  NONE: 'NONE',
  FOLLOWING: 'FOLLOWING',
  MUTUAL: 'MUTUAL',
  FOLLOW_BACK: 'FOLLOW_BACK',
} as const;